# backend/docs/swagger.yaml
basePath: /
definitions:
...
  controllers.CreateURLRequest:
    properties:
      expires_at:
        type: string
      url:
        type: string
    required:
    - url
    type: object
  controllers.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  controllers.RegisterRequest:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
...
  models.AuthLoginSuccess:
    properties:
      message:
        example: user logged in successfully
        type: string
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
  models.AuthRegisterSuccess:
    properties:
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
...
  models.CoreURLResponse:
    properties:
      short_code:
        type: string
      short_url:
        type: string
    type: object
...
  models.HTTPError:
    properties:
      error:
        example: invalid credentials
        type: string
    type: object
...
  models.MessageSuccess:
    properties:
      message:
        example: operation successful
        type: string
    type: object
  models.PaginatedURLResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.URL'
        type: array
      meta:
        $ref: '#/definitions/models.PaginationMeta'
    type: object
...
  models.PaginationMeta:
    properties:
      limit:
        example: 50
        type: integer
      page:
        example: 1
        type: integer
      total:
        example: 150
        type: integer
      total_pages:
        example: 3
        type: integer
    type: object
...
  models.Referrer:
    properties:
      clicks:
        type: integer
      source:
        type: string
    type: object
  models.Session:
    properties:
      browser:
        type: string
      device:
        type: string
      id:
        type: integer
      ip_address:
        type: string
      is_current_device:
        type: boolean
      last_active:
        type: string
      location:
        type: string
    type: object
  models.SessionListResponse:
    properties:
      sessions:
        items:
          $ref: '#/definitions/models.Session'
        type: array
    type: object
  models.URL:
    properties:
      clicks:
        type: integer
      created_at:
        type: string
      expiry:
        type: string
      id:
        description: go for uuid later
        type: integer
      long_url:
        type: string
      short_url:
        type: string
      user_id:
        type: integer
    required:
    - long_url
    - short_url
    type: object
...
  models.URLStats:
    properties:
      clicks:
        type: integer
      created_at:
        type: string
      expires_at:
        type: string
      last_accessed:
        type: string
      short_code:
        type: string
      top_referrers:
        items:
          $ref: '#/definitions/models.Referrer'
        type: array
    type: object
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      role:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: URL shortener with Redis caching, sliding-window limits, buffered analytics
    and Prometheus metrics.
  title: URL Shortener
  version: "1.0"
paths:
...
  /api/urls:
    post:
      consumes:
      - application/json
      description: Creates an anonymous URL; optional expires_at must be in the future.
      parameters:
      - description: URL to shorten
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.CreateURLRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.CoreURLResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.HTTPError'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/models.HTTPError'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/models.HTTPError'
      summary: Create a short URL
      tags:
      - urls
  /api/urls/{code}/stats:
    get:
      parameters:
      - description: Short code
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.URLStats'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.HTTPError'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/models.HTTPError'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/models.HTTPError'
      summary: Get statistics for a public short URL
      tags:
      - urls
...
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return a JWT token
      parameters:
      - description: Login Info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AuthLoginSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.HTTPError'
      summary: Login user
      tags:
      - auth
  /api/v1/auth/logout:
    post:
      description: Invalidate the user's JWT token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MessageSuccess'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.HTTPError'
      security:
      - Bearer: []
      summary: Logout user
      tags:
      - auth
  /api/v1/auth/me:
    get:
      description: Get current authenticated user info
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.HTTPError'
      security:
      - Bearer: []
      summary: Get current user info
      tags:
      - auth
  /api/v1/auth/refresh:
    post:
      parameters:
      - description: refresh_token cookie from login or registration
        in: header
        name: Cookie
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AuthLoginSuccess'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.HTTPError'
      summary: Rotate a refresh token and issue a new access token
      tags:
      - auth
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account; login separately for an access token
      parameters:
      - description: Registration Info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.AuthRegisterSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.HTTPError'
      summary: Register a new user
      tags:
      - auth
  /api/v1/auth/sessions:
    get:
      description: Get all active refresh token sessions for the current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SessionListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.HTTPError'
      security:
      - Bearer: []
      summary: Get active sessions
      tags:
      - auth
  /api/v1/auth/sessions/{id}:
    delete:
      description: Revokes a specific refresh token session for the current user
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MessageSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.HTTPError'
      security:
      - Bearer: []
      summary: Revoke a session
      tags:
      - auth
  /api/v1/auth/sessions/others:
    delete:
      description: Revokes all refresh token sessions for the current user except
        the one making this request
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MessageSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.HTTPError'
      security:
      - Bearer: []
      summary: Revoke all other sessions
      tags:
      - auth
...
  /api/v1/urls:
    get:
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 50
        description: Page size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PaginatedURLResponse'
      security:
      - Bearer: []
      summary: List the current user's short URLs
      tags:
      - legacy URLs
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"